Make dependencies installation faster
In requirements.txt, the versions were fixed. As an example, we had "wget==3.2"
which implies that pip downloaded only version 3.2 of wget. This leads
to dependcy bactracking issues, as stated in the following page:
https://pip.pypa.io/en/stable/topics/dependency-resolution/
Hence, as suggested in the above page, we simply change these versions
asumptions. In the example above, we replaced it by "wget>=3.2".